+++ src/com/vito/src/SmallPrimes.java
@@ -1,4 +1,4 @@
-package com.vito;
+package com.vito.src;
 
 import java.math.BigInteger;
 import java.util.List;
