+++ src/com/vito/test/PollardRhoTest.java
@@ -1,5 +1,6 @@
-package com.vito;
+package com.vito.test;
 
+import com.vito.src.PollardRho;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
